requirements.txt
requests==2.26.0
python-dotenv==0.19.0
flask==2.0.1
